refactor: Make runner::executor private
